Add separate class for enum objects
